+++ code/qcommon/cm_load_xbox.cpp
@@ -1224,8 +1224,8 @@ Writes the portal state to a savegame file
 */
 // having to proto this stuff again here is crap, but wtf?...
 //
-qboolean SG_Append(unsigned long chid, const void *data, int length);
-int SG_Read(unsigned long chid, void *pvAddress, int iLength, void **ppvAddressPtr = NULL);
+qboolean SG_Append(unsigned int chid, const void *data, int length);
+int SG_Read(unsigned int chid, void *pvAddress, int iLength, void **ppvAddressPtr = NULL);
 
 void CM_WritePortalState ()
 {
